qt: reparent video window to root whence UI closes
The video window has to exist until it is closed by its owner, i.e.
WindowClose() is called. If it stayed as a child of the main UI window,
it would be destroyed with the main UI window.

This reparents it (back) to the root window before the main UI window
gets destroyed. This works around #21875.

Gbp-Pq: Name 0006-qt-reparent-video-window-to-root-whence-UI-closes.patch