dav1d: limit the number of extra frames needed by the decoder
The i_extra_picture_buffers is used to add pictures to the pool that the core
will allocate. dav1d is actually using n_threads frames. And the core is
allocating 10 frames per default for AV1. So we need to add the missing ones.

(cherry picked from commit a32031dc0f5f32083fc54a21397bce732742ccbe) (rebased)
rebased:
- the code dav1d 1.0.0 in 3.0 uses different max versions

Signed-off-by: Steve Lhomme <robux4@ycbcr.xyz>

Gbp-Pq: Name 0004-dav1d-limit-the-number-of-extra-frames-needed-by-the.patch