fetcher: don't download network metadata if the user doesn't want to
The option is defined as "Allow metadata network access". If it's false (default but query during the first run) that means the
user doesn't want to check metadata from the network.
So we should not start a network job when there's a possibility to do it.

This is a regression from ea88b8d68e252445bcc305fd7a3dd45ef3ce0126.
Before that we only used FETCHER_SCOPE_LOCAL if the option wasn't set.

(cherry picked from commit 1a85d6793948c88afe35692d8c17a82089a58f33)

Gbp-Pq: Name 0079-fetcher-don-t-download-network-metadata-if-the-user-.patch